Add UI to the game
Added a basic UI to the game, consisting of the available choices
in the forms of buttons; a display of the choices from both the
player and the computer; and the display of the current score.

No winner is declared here, the game runs for how many rounds the
player wants.

The js code had to be totally rewritten, so comparisson of the
version will be necessary to see the differences.